#27 refactor : ImageToUri extension 함수 구현
