--- app/src/main/java/com/aoztg/greengrim/presentation/util/FloatToDpExtension.kt
@@ -1,8 +1,0 @@
-package com.aoztg.greengrim.presentation.util
-
-import android.content.res.Resources
-
-internal fun Float.toDp(): Float {
-    val density = Resources.getSystem().displayMetrics.density
-    return this / density
-}
